// src/app/modules/accounting/models/approve-loan-application.model.ts
export class ApproveSelectedLoanGetModel{
    AccFlag: number;
    AccTypeDesc: string;
    AtyClass: number;
    InstallmentAmount: number;
    LastInstallmentAmount: number;
    NoOfInstallment: number;
    PreviousPerformence: string;
    SacnAmount: number;
    StepOfLoan: number;
    TotalDeposit: number;
    FirstInstallmentDate: string;
    SecondInstallmentDate: string;
}
export class GroupCodeModel{
    Division: number;
    District: number;
    UpZila: number;
    Thana: number;
    AddressL1: string;
    AddressL2:string;
    AddressL3:string;
    RegNo: number;
    Regdesc: string;
    RegName: string;
    CheckValue:any[];
}